Make the `TreeChange` type truly generic across status and tree/tree diff
