+++ crates/but-core/src/unified_diff.rs
@@ -1,11 +1,11 @@
-use super::{worktree, UnifiedDiff};
+use super::{ChangeState, UnifiedDiff};
 use bstr::{BStr, BString, ByteSlice};
 use gix::diff::blob::platform::prepare_diff::Operation;
 use gix::diff::blob::unified_diff::ContextSize;
 use gix::diff::blob::ResourceKind;
 use serde::Serialize;
 
-/// A hunk as used in a [UnifiedDiff].
+/// A hunk as used in a [UnifiedDiff], which also contains all added and removed lines.
 #[derive(Debug, Clone, Serialize)]
 #[serde(rename_all = "camelCase")]
 pub struct DiffHunk {
@@ -49,8 +49,8 @@ impl UnifiedDiff {
         repo: &gix::Repository,
         path: &BStr,
         previous_path: Option<&BStr>,
-        current_state: impl Into<Option<worktree::ChangeState>>,
-        previous_state: impl Into<Option<worktree::ChangeState>>,
+        current_state: impl Into<Option<ChangeState>>,
+        previous_state: impl Into<Option<ChangeState>>,
         context_lines: u32,
     ) -> anyhow::Result<Self> {
         let current_state = current_state.into();
